Separate the presenters out, useful for the SDK
